py_tm_vault_client/requirements.txt
requests==2.22.0
python-dateutil==2.4.1
confluent-kafka==1.3.0
names==0.3.0
sshuttle==0.78.5
